refactor(periods)!: change eternity to int (#1229)
Previously, eternal periods and entities were populated with `inf`
values, which are floats. This was an exception, already, as the rest of
the values are integers. If you store, for example, thousands of
instants in a numpy array, just one eternal instant will force the whole
array to pass from integer to float. Now, eternal instants are populated
with `-1`, and can be produced with `Instant.eternity()` and
`Period.eternity()`. Also, they can be checked with `is_eternal`.