BREAKING CHANGE: remove CoreEntity.check_role_validity
The method in itself had no relation to `entities` (it was a static
method, so no depending neither on an entity class nor an instance). It
could have been kept as a `populations` helper, where it was exclusively
used. Nonetheless, `check_role_validity` had the peculiar behaviour of
not failing but when the value of `role` is not `None`, that granted it
in my opinion the right to be removed so as to avoid confusion.